Gap 16: nuint addressing in bulk memory ops + MemSlice chokepoint
Round 7 closed (int)ea truncation in the load/store helpers
(MemoryHelpers.{Load,Store}*) but missed the bulk-op family and
the [OpHandler]-dispatch chokepoint. Both had the same shape and
the same crash mode — any guest writing to a memory address past
int.MaxValue AVs the host process. Rust's release-mode
vec![0u8; N] lowers to a single memory.fill after cabi_realloc,
so non-trivial allocations past 2 GiB trip it.

Migrated to the nuint overloads round 7 added on MemoryInstance:

  - Wacs.Transpiler.Lib/AOT/Emitters/BulkEmitter.cs
    BulkHelpers.{MemoryCopy, MemoryFill, MemoryInit} — widen
    dst (and src for the dst-side memory in MemoryCopy) to
    nuint at the start, route through mem.AsSpan(nuint, int).
    MemoryInit's src stays int (data segment is byte[]-bounded).

  - Wacs.Core/Wacs.Core/Instructions/MemoryHandlers.cs MemSlice
    — single chokepoint for every [OpHandler] load/store
    dispatch. Last line `return mem.AsSpan((int)ea, width)`
    becomes `return mem.AsSpan((nuint)ea, width)`. This site
    was missed in round 7's per-instruction-file sweep.

  - Wacs.Core/Wacs.Core/Instructions/MemoryBulk.cs —
    InstMemoryInit.Execute, InstMemoryCopy.Execute,
    InstMemoryFill.Execute switch guest-memory address args
    from (int)x to (nuint)x.

Test surface: 3 new [Fact]s extend MemoryHelpersHighAddressTests:
MemoryFill_PastTwoGiB_FillsRegion, MemoryCopy_PastTwoGiB_CopiesRegion,
MemoryInit_PastTwoGiB_CopiesFromDataSegment. Each writes/copies
into a NativePointer 33000-page (~2.0625 GiB) memory at
addr = 0x80000400 (~2 GiB + 1 KiB) and asserts bytes round-trip.
Pre-fix each AVs; post-fix all complete cleanly. Total in the
suite: 8/8 (5 from gap 15 + 3 from gap 16).

Versions:
  - WACS                0.13.3 → 0.13.4
  - WACS.Cli            1.5.3  → 1.5.4
  - WACS.Transpiler.Lib 0.8.3  → 0.8.4

Test counts: Wacs.Core 394, Wacs.ComponentModel 354,
Wacs.Transpiler 767 (+3), Wacs.WASI.Preview2 189, Wacs.HostBindings
14, Spec.Test 770. All green. AOT smoke (wacs aot --wasi) still
produces a working native binary (5.8 MB osx-arm64, "hello"
output, exit 0).

Out of scope (separate gap): atomics. AtomicHelpers.CheckEa
returns int and the int ea cascades through the abstract
InstAtomicLoad.DoLoad / InstAtomicStore.DoStore signatures. Same
shape as gaps 15 + 16, different cohort (atomic-using shared-
memory threading guests). Follow-up.

+++ Wacs.Core/Wacs.Core/Instructions/MemoryBulk.cs
@@ -231,9 +231,12 @@ namespace Wacs.Core.Instructions
 
             // Bulk Span.CopyTo (memmove); mode-aware via mem.AsSpan
             // and zero-cost on byte[] (Span over the array's data).
+            // (nuint)d for the dst guest memory keeps the AsSpan
+            // dispatch in the unsigned range past 2 GiB. The data
+            // segment is byte[]-bounded so int casts on s/n are safe.
             if (n == 0) return;
             new ReadOnlySpan<byte>(data.Data, (int)s, (int)n)
-                .CopyTo(mem.AsSpan((int)d, (int)n));
+                .CopyTo(mem.AsSpan((nuint)d, (int)n));
         }
 
         public override InstructionBase Parse(BinaryReader reader)
@@ -384,9 +387,12 @@ namespace Wacs.Core.Instructions
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Destination memory overflow.");
 
             // Span<byte>.CopyTo dispatches to Buffer.Memmove and
-            // handles overlap correctly across both modes.
+            // handles overlap correctly across both modes. Both src
+            // and dst go through the (nuint, int) AsSpan overload so
+            // the high half of NativePointer memory addresses round-
+            // trip without sign truncation.
             if (n == 0) return;
-            memSrc.AsSpan((int)s, (int)n).CopyTo(memDst.AsSpan((int)d, (int)n));
+            memSrc.AsSpan((nuint)s, (int)n).CopyTo(memDst.AsSpan((nuint)d, (int)n));
         }
 
         public override InstructionBase Parse(BinaryReader reader)
@@ -454,9 +460,11 @@ namespace Wacs.Core.Instructions
                 || (ulong)mem.ByteLength - (ulong)d < (ulong)n)
                 throw new TrapException("Instruction memory.fill failed. Buffer overflow");
 
-            // Bulk Span.Fill — mode-aware via mem.AsSpan.
+            // Bulk Span.Fill — mode-aware via mem.AsSpan. The
+            // (nuint, int) overload preserves the full unsigned dst
+            // range so NativePointer memories past 2 GiB don't wrap.
             if (n == 0) return;
-            mem.AsSpan((int)d, (int)n).Fill((byte)(0xFF & cU32));
+            mem.AsSpan((nuint)d, (int)n).Fill((byte)(0xFF & cU32));
         }
 
         public override InstructionBase Parse(BinaryReader reader)
